requirements.txt
Flask==3.1.0
numpy==2.2.1
Pillow==11.0.0
Werkzeug==3.1.3
opencv-python-headless==4.2.0.32
gunicorn
